requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
timm>=0.9.12

# Medical imaging
nibabel>=5.2.0
SimpleITK>=2.3.1
monai>=1.3.0

# Augmentation
albumentations>=1.3.1

# Text encoder
open-clip-torch>=2.24.0
transformers>=4.36.0

# Report generation (Module 8)
peft>=0.7.0
bitsandbytes>=0.41

# Metrics & evaluation
torchmetrics>=1.2.0
rouge-score>=0.1.2

# Experiment tracking
wandb>=0.16.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
opencv-python>=4.8.0

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.0
pandas>=2.1.0
scipy>=1.11.0
scikit-image>=0.22.0
